fix: notication service
